backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
aiosqlite>=0.19.0
pydantic>=2.6.4
pyjwt>=2.10.1
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
qrcode[pil]>=7.4.2
pillow>=10.0.0
cryptography>=42.0.8
requests>=2.31.0
